add update contact action

+++ index.php
@@ -74,9 +74,9 @@ try {
                                         <td class="py-2 px-4 border-r border-stone-200 text-stone-500"><?= htmlspecialchars($contact['phone'] ?? '') ?></td>
                                         <td class="py-2 px-4 border-r border-stone-200 text-stone-500"><?= htmlspecialchars($contact['email']) ?></td>
                                         <td class="py-2 px-4 flex justify-between items-center">
-                                            <a href="#edit" class="text-neutral-600 hover:text-neutral-800 transition duration-150 ease-in-out">
+                                            <a href="form_contact.php?id=<?= $contact['id'] ?>" class="text-neutral-600 hover:text-neutral-800 transition duration-150 ease-in-out">
                                                 <i class="ph-duotone ph-pencil text-lg"></i></a>
-                                            <a href="actions/delete_contact.php?id=<?= $contact['id'] ?>" class="text-red-500 hover:text-red-600 transition duration-150 ease-in-out">
+                                            <a href="actions/delete_contact.php?id=<?= $contact['id'] ?>" onclick="return confirm('Êtes-vous sûr de vouloir supprimer ce contact ?')" class="text-red-500 hover:text-red-600 transition duration-150 ease-in-out">
                                                 <i class="ph-duotone ph-trash text-lg"></i></a>
                                         </td>
                                     </tr>
